Разделил код на модулиconst messages = [
